Add code nicely slipt into Vertex.hpp and Vertex.cpp
